+++ core/src/br/edu/utfpr/ShowDoMilhao.java
@@ -1,44 +1,26 @@
 package br.edu.utfpr;
 
-import com.badlogic.gdx.ApplicationAdapter;
-import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.*;
-import com.badlogic.gdx.scenes.scene2d.ui.Skin;
-import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
-import com.badlogic.gdx.scenes.scene2d.utils.Drawable;
-import com.badlogic.gdx.scenes.scene2d.utils.TiledDrawable;
-import com.badlogic.gdx.utils.ScreenUtils;
-import com.badlogic.gdx.ApplicationAdapter;
 import com.badlogic.gdx.Game;
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.assets.AssetManager;
-import com.badlogic.gdx.audio.Sound;
-import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.utils.ScreenUtils;
+import com.badlogic.gdx.graphics.g2d.TextureAtlas;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.InputListener;
 import com.badlogic.gdx.scenes.scene2d.Stage;
-import com.badlogic.gdx.scenes.scene2d.ui.Button;
-import com.badlogic.gdx.scenes.scene2d.ui.ImageButton;
 import com.badlogic.gdx.scenes.scene2d.ui.ImageTextButton;
 import com.badlogic.gdx.scenes.scene2d.ui.Label;
 import com.badlogic.gdx.scenes.scene2d.ui.Skin;
-import com.badlogic.gdx.scenes.scene2d.ui.TextButton;
 import com.badlogic.gdx.scenes.scene2d.utils.TextureRegionDrawable;
 import com.badlogic.gdx.utils.Align;
+import com.badlogic.gdx.utils.ScreenUtils;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 
 public class ShowDoMilhao extends Game {
     private AssetManager assetManager;
-    TextureAtlas mAtlas;
-    SpriteBatch batch;
-    Texture img;
 
-    private Stage stage;
-    private Label outputLabel;
 
     public void create() {
         assetManager = new AssetManager();
@@ -48,44 +30,6 @@ public class ShowDoMilhao extends Game {
 //		img = new Texture("imagens\\show-do-milhao.jpg");
 //		MainScreen mainScreen = new MainScreen();
         this.setScreen(new LoadingScreen(this));
-
-
-        batch = new SpriteBatch();
-        img = new Texture("imagens\\show-do-milhao.jpg");
-        stage = new Stage(new ScreenViewport());
-        Gdx.input.setInputProcessor(stage);
-        int Help_Guides = 12;
-        int row_height = Gdx.graphics.getWidth() / 12;
-        int col_width = Gdx.graphics.getWidth() / 12;
-
-
-        Skin mySkin = new Skin(Gdx.files.internal("skin/neon-ui.json"));
-        ImageTextButton button4 = new ImageTextButton("Jogar", mySkin);
-        button4.setSize(col_width * 4, (float) (row_height * 2));
-        button4.getStyle().imageUp = new TextureRegionDrawable(new TextureRegion(new Texture(Gdx.files.internal("imagens\\jogar.png"))));
-        button4.getStyle().imageDown = new TextureRegionDrawable(new TextureRegion(new Texture(Gdx.files.internal("imagens\\jogar.png"))));
-        button4.setPosition(col_width * 7, Gdx.graphics.getHeight() - row_height * 6);
-        button4.addListener(new InputListener() {
-            @Override
-            public void touchUp(InputEvent event, float x, float y, int pointer, int button) {
-                outputLabel.setText("JOGAR!");
-            }
-
-            @Override
-            public boolean touchDown(InputEvent event, float x, float y, int pointer, int button) {
-                outputLabel.setText("JOGAR!");
-                return true;
-            }
-        });
-        stage.addActor(button4);
-
-        outputLabel = new Label("JOGAR!", mySkin);
-        outputLabel.setSize(Gdx.graphics.getWidth(), row_height);
-        outputLabel.setPosition(0, row_height);
-        outputLabel.setAlignment(Align.center);
-        stage.addActor(outputLabel);
-
-
     }
 
 
@@ -104,12 +48,7 @@ public class ShowDoMilhao extends Game {
 ////		batch.draw(img, 0, 0);
 //		batch.end();
         super.render();
-        ScreenUtils.clear(0, 0, 0, 1);
-        batch.begin();
-        batch.draw(img, 0, 0);
-        batch.end();
-        stage.act();
-        stage.draw();
+
     }
 
     @Override
